Update to AGP 9.2.0
This is a huge API changing update, which required to update nearly
all dependencies

+++ core/smart/domain/src/main/java/com/buzbuz/smartautoclicker/core/domain/model/condition/ConditionMapper.kt
@@ -73,7 +73,7 @@ private fun TriggerCondition.OnCounterCountReached.toCounterReachedEntity(): Con
         counterName = counterName,
         counterComparisonOperation = comparisonOperation.toEntity(),
         counterOperationValueType = if (isNumberValue) CounterOperationValueType.NUMBER else CounterOperationValueType.COUNTER,
-        counterValue = if (isNumberValue) counterValue.value as Int else null,
+        counterValue = if (isNumberValue) counterValue.value else null,
         counterOperationCounterName = if (isNumberValue) null else counterValue.value as String,
         priority = 0,
     )
